+++ hello.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python3
-
-import sys
-print('Hello, Ish! This is Codey! 😊')
-
-# Add permission to execute the script
-import os
-os.chmod("./hello.py", 0o755)